Betting system created

+++ src/main/java/pl/coderslab/betting/entity/Role.java
@@ -4,6 +4,11 @@ import lombok.Data;
 
 import javax.persistence.*;
 
+/**
+ * This entity represents security role. It is described just by name(String name)
+ * There are only two roles: "ROLE_USER" and "ROLE_ADMIN".
+ */
+
 @Entity
 @Data
 @Table(name = "roles")
@@ -11,7 +16,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private int id;
+    private Long id;
     @Column(name = "role")
     private String name;
 }
